Draw X and O by Taking Turns
Combine Draw X and O functions by counting turn number to decide which symbol to draw. If the turn is an odd number, then X will be drawn (Therefore, X will always go first.) and, if the turn is an even number, O will be drawn while turn count will not exceed 9.

The programme will display turn count and whose turn is it. If the turn count exceeds 9. It will display Game Over.

Error Encounter : If the user clicks on the same grid, the programme will draw a new symbol on top of it.