Fix Commit 6's Error
Create a flat array of 9 elements of null to represent the state of each grids by null meaning the grid is empty. Then, create a variable to track the symbol of the player in their turn using the same logic in Commit 6. 

When the grid is clicked, the program will check the 9x array if the grid is empty or not. If empty, the program will compute the symbol of the player before drawing it and, then, update the element of the array from null to that symbol in the index according to that grid.